requirements.txt
# Core AI/ML Frameworks
openai==1.54.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
langgraph==0.2.45
crewai==0.80.0
crewai-tools==0.14.0

# Web Automation & APIs
requests==2.32.3
httpx==0.27.2
aiohttp==3.10.10
playwright==1.48.0
selenium==4.26.1

# Social Media APIs
instagrapi==2.1.2
tweepy==4.14.0
python-linkedin-v2==0.1.0
facebook-sdk==3.1.0

# Data Processing
pandas==2.2.3
numpy==2.1.3
python-dateutil==2.9.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pymongo==4.10.1

# Google Integration
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gspread==6.1.4
oauth2client==4.1.3

# Image Processing
Pillow==11.0.0
canva-api==0.1.0

# Scheduling & Task Management
schedule==1.2.2
celery==5.4.0
redis==5.2.0

# Analytics & Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic==2.10.2
pydantic-settings==2.6.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
python-slugify==8.0.4
pytz==2024.2
tqdm==4.67.1
validators==0.34.0

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==2.18.0

# API Documentation
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.2

# Async Support
asyncio==3.4.3
nest-asyncio==1.6.0
